+++ build.gradle
@@ -23,6 +23,12 @@ dependencies {
 	//Apache Commons Lang
 	implementation 'org.apache.commons:commons-lang3:3.14.0'
 
+	//MySQL Driver
+	runtimeOnly 'com.mysql:mysql-connector-j'
+
+	//MyBatis
+	implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter:3.0.3'
+
 
 	providedRuntime 'org.springframework.boot:spring-boot-starter-tomcat'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
