+++ src/main/java/raisetech/StudentManagement/Student.java
@@ -1,7 +1,6 @@
 package raisetech.StudentManagement;
 
 public class Student {
-
   private String name;
   private int age;
