+++ src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=StudentManagement
+
+spring.datasource.url=jdbc:mysql://localhost:3306/StudentManagement
+spring.datasource.username=root
+spring.datasource.password=pokepoke929
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+mybatis.configuration.map-underscore-to-camel-case=true
